+++ js/app/actions/UserActions.js
@@ -1,7 +1,39 @@
-var KanbanerDispatcher = require('../dispatcher/KanbanerDispatcher');
-var KanbanerConstants = require('../constants/KanbanerConstants');
+var
+  Q = require( "q" ),
+  $ = require( "jquery" ),
+  btoa = require( "btoa" ),
 
-var ActionTypes = KanbanerConstants.ActionTypes;
+  KanbanerDispatcher = require( "../dispatcher/KanbanerDispatcher" ),
+  KanbanerConstants = require( "../constants/KanbanerConstants" ),
+
+  ActionTypes = KanbanerConstants.ActionTypes;
+
+/**
+ * Creates bases authorization token. sends it to github.com.
+ * If authorization success, send object with user and token to next promise ,
+ * Else send null to next promise
+ * @param {string} username
+ * @param {string} password
+ * @return {object} promise object
+ */
+var authenticateUser = function( username, password ) {
+  var authToken = btoa( username + ":" + password );
+  return Q.when($.ajax({
+    url: "https://api.github.com/user",
+    method: "get",
+    headers: {
+      "Authorization": "Basic " + authToken
+    }
+  }) )
+    .then(function ( data ) {
+      return Q({
+        user: data,
+        token: authToken
+      });
+    }, function() {
+      return Q( null );
+    });
+};
 
 var UserActions = {
   /**
@@ -9,16 +41,30 @@ var UserActions = {
    * @param {string} password
    */
   sendAuthCredentials: function(username, password) {
-    KanbanerDispatcher.handleViewAction({
-      type: ActionTypes.SEND_AUTHENTICATION_CREDENTIALS,
-      username: username,
-      password: password
-    });
+    return authenticateUser(username, password)
+      .then(function( data ) {
+        if( data ) {
+          KanbanerDispatcher.handleViewAction({
+            type: ActionTypes.USER_LOGIN_SUCCESS,
+            user: data.user,
+            token: data.token
+          });
+          return Q( "succcess" );
+        } else {
+          KanbanerDispatcher.handleViewAction({
+            type: ActionTypes.USER_LOGIN_FAIL
+          });
+          return Q( "failed" );
+        }
+      })
   },
 
   logoutUser: function() {
-    KanbanerDispatcher.handleViewAction({
-      type: ActionTypes.USER_LOGOUT
+    return Q.fcall(function() {
+      KanbanerDispatcher.handleViewAction({
+        type: ActionTypes.USER_LOGOUT
+      });
+      return "logout";
     });
   }
 };